Encoder Modular

+++ PIDmotores/BNO.cpp
@@ -7,8 +7,8 @@
 #include "BNO.h"
 
 BNO::BNO(){
-    this-> event = {0};
-    this-> bno = Adafruit_BNO055(55, 0x28, &Wire);
+    this->event = {0};
+    this->bno = Adafruit_BNO055(55, 0x28, &Wire);
 }
 void BNO::setupBNO(){
     adafruit_bno055_opmode_t mode = OPERATION_MODE_IMUPLUS;
@@ -16,8 +16,7 @@ void BNO::setupBNO(){
     Serial.println("Orientation Sensor Test"); Serial.println("");
 
     // Initialise the sensor 
-    if (!bno.begin())
-    {
+    if (!bno.begin()){
         // There was a problem detecting the BNO055 ... check your connections
         Serial.print("Ooops, no BNO055 detected ... Check your wiring or I2C ADDR!");
         while(1);
@@ -30,7 +29,7 @@ void BNO::setupBNO(){
 void BNO::updateBNO(sensors_event_t &event){
     bno.getEvent(&event);
 }
-float BNO::getOrientationX(){
+double BNO::getOrientationX(){
     updateBNO(event);
     return event.orientation.x;
 }
